// frontend/src/lib/services/backend-api-generated/models/sub-category.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bakebee Admin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface SubCategory {
    'id'?: string;
    'sub_category_name'?: string;
    'main_category_id'?: string;
    'is_active'?: boolean;
}